+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Create: Integrated Farming 1.2.6
 
-### Add
+### Update
 * Added Create Harvester support for Twilight's Flavor & Delight Lily Pad farming.
+* Added Ponder scene for My Nether's Delight Powdery Crop harvesting with Mechanical Arms.